+++ test/test_helper.rb
@@ -88,6 +88,10 @@ ActiveRecord::Schema.define do
   end
 end
 
+class ApplicationSystemTestCase < ActionDispatch::SystemTestCase
+  driven_by :selenium, using: :headless_chrome, screen_size: [1200, 900]
+end
+
 module ActiveSupport
   class TestCase
     # This suite does not use transactional tests, and that is deliberate.